Link Layer only process FRAMEs now.
The code has been significantly restructured to make the link layer and
transport layer more distinct in terms of their functionality. By
letting the link layer only focusing with one datatype, this makes
optimization and processing a lot easier and faster.

-FRAME is basically the old UMPACKET.
-Packets are now referred to as transport layer objects. UMPACKET is
aliased as a FRAME in the transport layer.
-Fields needed by USPACKET and RSPACKET are now part of the frame's
payload.

-Packet.c has been empty. It will be reimplemented properly afterwards.